+++ src/modules/patient/patient.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { PatientController } from './patient.controller';
-import { PatientService } from './patient.service';
+import { PatientService } from './service/patient.service';
 import { MulterModule } from '@nestjs/platform-express';
 import { ExcelParserService } from './service/excel-parser.service';
 import { PatientRepositoryModule } from '@domain/patient/patient-repository.module';
